a bit of refactoring
